add: floating-point operations

+++ simulator/util.hpp
@@ -5,8 +5,8 @@
 unsigned int id_of_pc(unsigned int n);
 int read_reg(int i);
 void write_reg(int i, int v);
-int read_reg_fp(int i);
-void write_reg_fp(int i, int v);
+float read_reg_fp(int i);
+void write_reg_fp(int i, float v);
 int binary_stoi(std::string s);
 std::string string_of_op(Operation &op);
 void print_reg();
